major bug fixes
pending: move out buttons from functions and sale options

+++ POS/src/com/malabon/object/Recipe.java
@@ -7,5 +7,5 @@ public class Recipe implements Serializable {
 	public int recipe_id;
 	public int product_id;
 	public int ingredient_id;
-	public int ingredient_qty;
+	public double ingredient_qty;
 }
